Lots of cleanup and renames.

+++ make_darknet_cjdns.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 
-package_list="radvd nodejs build-essential python2.7"
+package_list="nodejs build-essential python2.7"
 
 ###########
 # Install #
@@ -15,17 +15,23 @@ for package in $package_list; do
       new_packages="$new_packages $package"
    fi
 done
-echo $new_packages > ./new_packages_darknet.txt
+echo $new_packages > ./new_packages_darknet_cjdns.txt
 
-# Build cjdns
-wget -O cjdns.tar.gz https://github.com/cjdelisle/cjdns/archive/master.tar.gz
-tar xzf cjdns.tar.gz
-cd cjdns-master
-./do
-sudo mv cjdroute /usr/bin/
-sudo mv contrib/systemd/cjdns.service /etc/systemd/system/
-sudo systemctl enable cjdns
-sudo systemctl start cjdns
+# Dont install if we already have a tun0. Just do everything around it.
+if [ ! -f /sys/class/net/tun0 ]; then
+   # Build cjdns
+   wget -O cjdns.tar.gz https://github.com/cjdelisle/cjdns/archive/master.tar.gz
+   tar xzf cjdns.tar.gz
+   cd cjdns-master
+   ./do
+   sudo chown root:root cjdroute contrib/systemd/cjdns.service
+   sudo mv cjdroute /usr/bin/
+   sudo mv contrib/systemd/cjdns.service /etc/systemd/system/
+   cd ..
+   sudo rm -rf cjdns-master
+   sudo systemctl enable cjdns
+   sudo systemctl start cjdns
+fi
 
 # IP Tables; allow tap0 to talk to the outside via ethernet
 sudo cp /etc/iptables/rules.v6 /etc/iptables/rules.v6_back
@@ -36,28 +42,6 @@ sudo sh -c '( ip6tables-save > /etc/iptables/rules.v6 )'
 
 sudo sh -c '( echo "net.ipv6.conf.all.forwarding=1
 net.ipv6.ip_forward = 1" > /etc/sysctl.d/belagos-darknet-sysctl.conf )'
-
-# Add "fdfc::1" address as static on tap0
-sudo sh -c "( echo \"
-iface tap0 inet6 static
-   address fdfc::1
-   netmask 64\" >> /etc/network/interfaces.d/tap0.iface )"
-
-# radvd, but it doesnt seem like it works in plan 9. But it does on OpenBSD.
-sudo sh -c "( echo \"
-interface tap0
-{
-	AdvSendAdvert on;
-	prefix fdfc::1/64 {
-		AdvRouterAddr on;
-	};
-};\" > /etc/radvd.conf )"
-sudo systemctl enable radvd
-sudo systemctl restart radvd
-
-# Final step needed as radvd doesnt seem to work...
-echo "Run the following on your plan 9 machine:"
-echo "echo 'add :: 0:0:0:0:0:0:0:0 fdfc::1' >/net/iproute"
 }
 
 #############
@@ -68,17 +52,18 @@ uninstall()
 sudo mv /etc/iptables/rules.v6_back /etc/iptables/rules.v6
 
 # Remove Packages
-sudo DEBIAN_FRONTEND=noninteractive apt-get remove --purge -y `cat ./new_packages_darknet.txt`
+sudo DEBIAN_FRONTEND=noninteractive apt-get remove --purge -y `cat ./new_packages_darknet_cjdns.txt`
 rm ./new_packages.txt
 sudo apt-get autoremove -y
 sudo apt-get clean -y
 
-sudo rm /etc/sysctl.d/belagos-darknet-sysctl.conf
-sudo rm /etc/radvd.conf
+# Remove CJDNS
+sudo systemctl stop cjdns
+sudo systemctl disable cjdns
+sudo rm /usr/bin/cjdroute
+sudo rm /etc/systemd/system/cjdns.service
 
-# Final step needed as radvd doesnt seem to work...
-echo "Run the following on your plan 9 machine:"
-echo "echo 'delete :: 0:0:0:0:0:0:0:0' >/net/iproute"
+sudo rm /etc/sysctl.d/belagos-darknet-sysctl.conf
 }
 
 ########
